requirements.txt
# HTTP client
httpx>=0.27.0
aiohttp>=3.9.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# Scheduling
apscheduler>=3.10.0

# Configuration
python-dotenv>=1.0.0

# Logging
python-json-logger>=2.0.0

# LLM client
openai>=1.0.0

# UI
streamlit>=1.28.0
streamlit-folium>=0.15.0
pandas>=2.0.0
folium>=0.14.0

# Utilities
python-dateutil>=2.8.0
